backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
pydantic>=2.5.0
python-multipart>=0.0.6
